@@ -6,19 +6,26 @@
       "ByJG\\Cache\\": "src/"
     }
   },
+  "autoload-dev": {
+    "psr-4": {
+      "Tests\\": "tests/"
+    }
+  },
   "require": {
-    "php": ">=7.4",
-    "psr/cache": "^1.0|^2.0",
+    "php": ">=8.1 <8.4",
+    "psr/cache": "^1.0|^2.0|^3.0",
     "psr/log": "^1.0|^1.1|^2.0",
     "psr/simple-cache": "^1.0|^2.0",
     "psr/container": "^1.0|^1.1|^2.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "5.7.*|7.4.*|^9.5"
+    "phpunit/phpunit": "^9.6",
+    "vimeo/psalm": "^5.9"
   },
   "suggest": {
     "ext-memcached": "*",
-    "ext-redis": "*"
+    "ext-redis": "*",
+    "ext-shmop": "*"
   },
   "provide": {
     "psr/cache-implementation": "1.0",
