@@ -3,27 +3,37 @@
 namespace ByJG\Cache\Psr16;
 
 use ByJG\Cache\CacheLockInterface;
+use ByJG\Cache\Exception\InvalidArgumentException;
+use DateInterval;
+use Psr\Container\ContainerExceptionInterface;
+use Psr\Container\NotFoundExceptionInterface;
 
 class NoCacheEngine extends BaseCacheEngine implements CacheLockInterface
 {
     /**
-     * @param string $key The object KEY
-     * @param int $default IGNORED IN MEMCACHED.
-     * @return mixed Description
+     * @param string $key
+     * @param mixed $default
+     * @return mixed
+     * @throws ContainerExceptionInterface
+     * @throws InvalidArgumentException
+     * @throws NotFoundExceptionInterface
      */
-    public function get($key, $default = null)
+    public function get(string $key, mixed $default = null): mixed
     {
         $key = $this->getKeyFromContainer($key);
         return $default;
     }
 
     /**
-     * @param string $key The object Key
-     * @param object $value The object to be cached
-     * @param int $ttl The time to live in seconds of this objects
-     * @return bool If the object is successfully posted
+     * @param string $key
+     * @param mixed $value
+     * @param DateInterval|int|null $ttl
+     * @return bool
+     * @throws ContainerExceptionInterface
+     * @throws InvalidArgumentException
+     * @throws NotFoundExceptionInterface
      */
-    public function set($key, $value, $ttl = 0)
+    public function set(string $key, mixed $value, DateInterval|int|null $ttl = null): bool
     {
         $key = $this->getKeyFromContainer($key);
         return true;
@@ -32,8 +42,11 @@ class NoCacheEngine extends BaseCacheEngine implements CacheLockInterface
     /**
      * @param string $key
      * @return bool
+     * @throws ContainerExceptionInterface
+     * @throws InvalidArgumentException
+     * @throws NotFoundExceptionInterface
      */
-    public function delete($key)
+    public function delete(string $key): bool
     {
         $key = $this->getKeyFromContainer($key);
         return true;
@@ -43,7 +56,7 @@ class NoCacheEngine extends BaseCacheEngine implements CacheLockInterface
      * Lock resource before set it.
      * @param string $key
      */
-    public function lock($key)
+    public function lock(string $key): void
     {
         return;
     }
@@ -52,12 +65,12 @@ class NoCacheEngine extends BaseCacheEngine implements CacheLockInterface
      * UnLock resource after set it
      * @param string $key
      */
-    public function unlock($key)
+    public function unlock(string $key): void
     {
         return;
     }
 
-    public function isAvailable()
+    public function isAvailable(): bool
     {
         return true;
     }
@@ -67,7 +80,7 @@ class NoCacheEngine extends BaseCacheEngine implements CacheLockInterface
      *
      * @return bool True on success and false on failure.
      */
-    public function clear()
+    public function clear(): bool
     {
         return true;
     }
@@ -81,10 +94,11 @@ class NoCacheEngine extends BaseCacheEngine implements CacheLockInterface
      *
      * @param string $key The cache item key.
      * @return bool
-     * @throws \Psr\SimpleCache\InvalidArgumentException
-     *   MUST be thrown if the $key string is not a legal value.
-    */
-    public function has($key)
+     * @throws ContainerExceptionInterface
+     * @throws InvalidArgumentException
+     * @throws NotFoundExceptionInterface
+     */
+    public function has(string $key): bool
     {
         $key = $this->getKeyFromContainer($key);
         return false;
